// src/server/api/models/investing/models/FundTransactionRequest.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * A JSON blob representing a fund transaction request. The fields marked with an * means that they are required.
 */
export type FundTransactionRequest = {
    /**
     * A unique identifier for the investment account that is generated by the investing service.
     */
    accountId: string;
    /**
     * Indicates the order side ie SUBSCRIPTION | REDEMPTION .
     */
    orderType: FundTransactionRequest.orderType;
    /**
     * The market || execution venue identifier for the order
     */
    marketCode: string;
    /**
     * The ID for the security to be traded.
     */
    secId: string;
    /**
     * The quantity requested. This is validated using the security master instrument definition. If this is not set then a valid transaction value must be provided. If this is set, it will be used to calculate the transaction value.
     */
    requestedQty?: number;
    /**
     * This is optional and if the requested quantity is not set, it will be used to calculate the transaction quantity
     */
    requestedVal?: number;
    /**
     * Optionally specify a price to be used for the transaction. If this is not provided, the bid / offer price for the transaction date will be used.
     */
    price?: number;
    /**
     * The request date
     */
    requestDate: string;
    /**
     * The specified value must match the currency configured in the instrument definition.
     */
    currency: string;
    /**
     * Notes related to the record
     */
    notes?: string;
    /**
     * Indicates if the order should be auto approved
     */
    autoApprove: boolean;
    /**
     * After redeeming units and this flag is true, the cash will be automatically paid out. If the flag is false, the cash will be left in the investment cash account and can be reinvested. The default value is false if it is not specified.
     */
    autoPayRedemption?: boolean;
    /**
     * The channel through which the transaction was submitted
     */
    channel?: FundTransactionRequest.channel;
};

export namespace FundTransactionRequest {

    /**
     * Indicates the order side ie SUBSCRIPTION | REDEMPTION .
     */
    export enum orderType {
        SUBSCRIPTION = 'SUBSCRIPTION',
        REDEMPTION = 'REDEMPTION',
    }

    /**
     * The channel through which the transaction was submitted
     */
    export enum channel {
        NA = 'NA',
        WEB = 'WEB',
        MOBILE = 'MOBILE',
        USSD = 'USSD',
        ATM = 'ATM',
        POS = 'POS',
        BRANCH = 'BRANCH',
        EXEC_REPORT = 'EXEC_REPORT',
        FILE_UPLOAD = 'FILE_UPLOAD',
        API = 'API',
        EMAIL = 'EMAIL',
        SMS = 'SMS',
        IVR = 'IVR',
        KIOSK = 'KIOSK',
        CALL_CENTER = 'CALL_CENTER',
        OTHER = 'OTHER',
    }


}